frontend/requirements.txt
streamlit>=1.30.0
requests>=2.25.0
Pillow>=9.0.0
urllib3>=1.26.0